consumer/requirements.txt
river==0.11.1
bytewax==0.9.0
kafka-python